+++ src/routines/level2/xtpmv.cpp
@@ -10,7 +10,6 @@
 #include "routines/level2/xtpmv.hpp"
 
 #include <string>
-#include <vector>
 
 namespace clblast {
 // =================================================================================================
@@ -32,8 +31,8 @@ void Xtpmv<T>::DoTpmv(const Layout layout, const Triangle triangle, const Transp
   x_buffer.CopyTo(queue_, x_size, scratch_buffer);
 
   // The data is either in the upper or lower triangle
-  size_t is_upper = ((triangle == Triangle::kUpper && layout != Layout::kRowMajor) ||
-                     (triangle == Triangle::kLower && layout == Layout::kRowMajor));
+  size_t is_upper = static_cast<size_t>((triangle == Triangle::kUpper && layout != Layout::kRowMajor) ||
+                                        (triangle == Triangle::kLower && layout == Layout::kRowMajor));
 
   // Adds '2' to the parameter if the diagonal is unit
   auto parameter = (diagonal == Diagonal::kUnit) ? is_upper + 2 : is_upper;
